refactor: 포인트 퍼센트 전략 인터페이스 분리
