+++ src/main/java/cart/domain/point/PointPolicyStrategy.java
@@ -1,0 +1,6 @@
+package cart.domain.point;
+
+public interface PointPolicyStrategy {
+    Long caclulatePointWithPolicy(Long amountPayment);
+    int getPointPercentage();
+}
